requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
openai==1.6.1
requests==2.31.0
pillow==10.1.0
